edit

+++ lib/tasks/process_deal_coupang.rb
@@ -7,10 +7,16 @@ browser = Watir::Browser.new
   while 1
     search_key = DealSearchWord.all.pluck(:word)
     p "쿠팡"
-    DealItem.add_coupang(browser, search_key)
+    ret = DealItem.add_coupang(browser, search_key)
+    unless ret
+      browser.close
+      headless.destroy
+      
+      headless = Headless.new
+      headless.start
+      browser = Watir::Browser.new
+    end
   end
-browser.close
-headless.destroy
