+++ lib/tasks/process_deal_g9.rb
@@ -8,11 +8,16 @@ browser = Watir::Browser.new
     search_key = DealSearchWord.all.pluck(:word)
     
     #G9
-    p "G9"
-    DealItem.add_g9(browser, search_key)
+    ret = DealItem.add_g9(browser, search_key)
+    unless ret
+      browser.close
+      headless.destroy
+      
+      headless = Headless.new
+      headless.start
+      browser = Watir::Browser.new
+    end
   end
-browser.close
-headless.destroy
